+++ Projects/Engine/Source/Asset/GLTFAssetImporter.cpp
@@ -514,7 +514,7 @@ namespace Odyssey
 		{
 			FBXBone& bone = m_RigData.Bones[node->name];
 			bone.Name = node->name;
-			bone.Index = m_RigData.BoneCount;
+			bone.Index = (int32_t)m_RigData.BoneCount;
 			bone.NodeIndex = nodeIndex;
 			m_RigData.BoneCount++;
 
@@ -740,7 +740,7 @@ namespace Odyssey
 				}
 			}
 
-			size_t maxFrames = std::ceil(m_AnimationData.Duration * (double)m_AnimationData.FramesPerSecond);
+			size_t maxFrames = (size_t)std::ceil(m_AnimationData.Duration * (double)m_AnimationData.FramesPerSecond);
 			double step = 1.0 / (double)m_AnimationData.FramesPerSecond;
 
 			for (auto& [boneName, boneKeyframe] : m_AnimationData.BoneKeyframes)
